feat(viewer): mirror the display path into the WASM viewer (parity)
The display-mode query + binary encoding was server-only (src/viewer/routes.rs),
so the static WASM viewer — which shares the frontend but has its own backend —
silently fell back to the un-decimated JSON path.

Move the reducer options + binary wire encoding into the shared `dashboard`
crate (`display_wire`) and call it from both shells:
- server `range_query_display` now delegates to `display_wire::display_query`
- WASM `Viewer`/`WasmCaptureRegistry` gain `query_range_display`, returning the
  same binary body (byte-identical by construction)
- the WASM copy of `viewer_api.js` gains `queryRangeDisplay` wired to it

Verified: server path passes tests/display_synthetic.sh (shared encoder), and
the WASM crate compiles for wasm32 with the display method. Rebuilding the WASM
pkg (crates/viewer/build.sh) picks it up for the static site.
